pyproject.toml
[tool.poetry]
name = "fasttowertestproject"
version = "0.1.0"
description = ""
authors = ["Skelesheit <Skelesheit@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fasttower = {path = "fasttower-0.1.4-py3-none-any.whl"}
sqlalchemy = "^2.0.36"
sqladmin = "^0.20.1"
pytest = "^8.3.3"
httpx = "^0.27.2"
alembic = "^1.14.0"
sqlmodel = "^0.0.22"
pydantic = {extras = ["email"], version = "^2.9.2"}
pyjwt = "^2.9.0"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
